Changing server and metricRouter to typescript
