+++ tsconfig.json
@@ -1,0 +1,27 @@
+{
+    "compilerOptions": {
+      "target": "es5",                 // Compile to ES5
+      "lib": ["dom", "dom.iterable", "esnext"], // Include DOM and ESNext lib for frontend code
+      "allowJs": true,                 // Allow JavaScript files to be compiled
+      "skipLibCheck": true,            // Skip type checking of all declaration files (*.d.ts)
+      "esModuleInterop": true,         // Enables emit interoperability between CommonJS and ES Modules
+      "allowSyntheticDefaultImports": true, // Allow default imports from modules with no default export
+      "strict": true,                  // Enable all strict type-checking options
+      "forceConsistentCasingInFileNames": true, // Disallow inconsistently-cased references to the same file
+      "module": "commonjs",            // Use commonjs module resolution
+      "moduleResolution": "node",      // Use Node.js module resolution strategy
+      "resolveJsonModule": true,       // Allow importing .json files
+      "isolatedModules": true,         // Ensure each file can be safely transpiled without relying on other imports
+      "noEmit": true,                  // Do not emit outputs (useful when using Babel for transpilation)
+      "jsx": "react-jsx"               // Support JSX in .tsx files
+    },
+    "include": [
+      "client/src/**/*", // Include your frontend source files
+      "server/**/*"      // Include your backend source files
+    ],
+    "exclude": [
+      "node_modules",    // Exclude the node_modules directory
+      "**/*.spec.ts"     // Exclude test files
+    ]
+}
+
